Fix range violation
If `auto` is used in a for loop for the count variable, `.length` is
used as the initializer and the counter is counting towards 0, `i--`.
Then the count variable will be of type `size_t`, since that's the
type of `.length`. This will cause the loop condition to never be
false because the `size_t` variable will wrap around to `size_t.max`
instead of `-1`.

https://github.com/d-widget-toolkit/dwt/issues/24